PreferencesWindow: Convert MaxValue to long for checking invalid time

+++ Apollo/Windows/PreferencesWindow.cs
@@ -108,7 +108,7 @@ namespace Apollo.Windows {
         void UpdateTime(object sender, EventArgs e) {
             CurrentSession.Text = $"Current session: {Program.TimeSpent.Elapsed.Humanize(minUnit: TimeUnit.Second, maxUnit: TimeUnit.Hour)}";
 
-            if (Preferences.Time >= (int)TimeSpan.MaxValue.TotalSeconds) Preferences.BaseTime = 0;
+            if (Preferences.Time >= (long)TimeSpan.MaxValue.TotalSeconds) Preferences.BaseTime = 0;
 
             AllTime.Text = $"All time: {Preferences.Time.Seconds().Humanize(minUnit: TimeUnit.Second, maxUnit: TimeUnit.Hour)}";
         }
